Introduce alembic for type conversions, and correct the datatypes.
